requirements.txt
langchain==0.2.16
streamlit==1.52.2
pandas==2.3.3
numpy==1.26.4
sentence-transformer==5.2.0
pydantic==2.12.5
torch==2.9.1
openpyxl==3.1.5
plotly==6.5.2